Initial version of the buildpack for testing
This is based on the jemalloc buildpack but modified to install mimalloc
instead. In the near term jemalloc should continue to work well but the
long term future is uncertain[1]. Given that is seems worthwhile to
consider other options and I've read good things about mimalloc and it
recently became an option for python.

This commit includes an initial classic buildpack implementation as well
as a partial CNB version. Without access to Fir to finalize the CNB
version a lot of it is just guesswork and likely broken.

1: https://jasone.github.io/2025/06/12/jemalloc-postmortem/